imported database; fixed bugs
