+++ package.json
@@ -9,7 +9,7 @@
   },
   "devDependencies": {},
   "scripts": {
-    "start": " node src/index.js"
+    "start": "node src/index.js"
   },
   "repository": {
     "type": "git",
